nguyenpham changed

+++ oganicshop/src/main/java/com/nguyenpham/oganicshop/dto/ProductResponse.java
@@ -23,12 +23,11 @@ public class ProductResponse {
     private int price;
     private int discount;
     private int finalPrice;
-    private int rating;
+    private double rating;
     private int numberOfReviews;
     private int amount;
     private int amountSold;
     private String status;
     private String categoryName;
-//    private String supplierName;
     private List<ReviewResponse> reviews;
 }
